+++ Box_Smasher/Assets/Animations/Player/Bodys.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c13ac4ce3ccee8b49b005a2c3039706b
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
