+++ Box_Smasher/Assets/Animations/Player/Faces/Charging.anim
@@ -20,20 +20,18 @@ AnimationClip:
   m_PPtrCurves:
   - curve:
     - time: 0
-      value: {fileID: 21300000, guid: 6ae6b004375497b4f8b2e98b683a7851, type: 3}
-    - time: 1
       value: {fileID: 21300000, guid: 12b2d96515a722549af867597cb047a2, type: 3}
-    - time: 2
+    - time: 0.25
       value: {fileID: 21300000, guid: 15fd878d312b2d14691b39fab2505a74, type: 3}
-    - time: 3
+    - time: 0.5
       value: {fileID: 21300000, guid: 62c01f096d1e5aa49b636932065e2620, type: 3}
-    - time: 4
+    - time: 0.75
       value: {fileID: 21300000, guid: 53b1ff8ff17de1e498c7828f10752122, type: 3}
     attribute: m_Sprite
     path: 
     classID: 212
     script: {fileID: 0}
-  m_SampleRate: 1
+  m_SampleRate: 4
   m_WrapMode: 0
   m_Bounds:
     m_Center: {x: 0, y: 0, z: 0}
@@ -48,7 +46,6 @@ AnimationClip:
       customType: 23
       isPPtrCurve: 1
     pptrCurveMapping:
-    - {fileID: 21300000, guid: 6ae6b004375497b4f8b2e98b683a7851, type: 3}
     - {fileID: 21300000, guid: 12b2d96515a722549af867597cb047a2, type: 3}
     - {fileID: 21300000, guid: 15fd878d312b2d14691b39fab2505a74, type: 3}
     - {fileID: 21300000, guid: 62c01f096d1e5aa49b636932065e2620, type: 3}
@@ -58,7 +55,7 @@ AnimationClip:
     m_AdditiveReferencePoseClip: {fileID: 0}
     m_AdditiveReferencePoseTime: 0
     m_StartTime: 0
-    m_StopTime: 5
+    m_StopTime: 1
     m_OrientationOffsetY: 0
     m_Level: 0
     m_CycleOffset: 0
