+++ Box_Smasher/Assets/Animations/Player/Faces/Falling.anim
@@ -20,8 +20,6 @@ AnimationClip:
   m_PPtrCurves:
   - curve:
     - time: 0
-      value: {fileID: 21300000, guid: 6ae6b004375497b4f8b2e98b683a7851, type: 3}
-    - time: 1
       value: {fileID: 21300000, guid: bfb67af6688cd4d4ca9f05b20aafdb39, type: 3}
     attribute: m_Sprite
     path: 
@@ -42,14 +40,13 @@ AnimationClip:
       customType: 23
       isPPtrCurve: 1
     pptrCurveMapping:
-    - {fileID: 21300000, guid: 6ae6b004375497b4f8b2e98b683a7851, type: 3}
     - {fileID: 21300000, guid: bfb67af6688cd4d4ca9f05b20aafdb39, type: 3}
   m_AnimationClipSettings:
     serializedVersion: 2
     m_AdditiveReferencePoseClip: {fileID: 0}
     m_AdditiveReferencePoseTime: 0
     m_StartTime: 0
-    m_StopTime: 2
+    m_StopTime: 1
     m_OrientationOffsetY: 0
     m_Level: 0
     m_CycleOffset: 0
