+++ Box_Smasher/Assets/Prefabs/BS/BS_Boss.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: bbd7f9c547c6f0a4587f5c67781702ae
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
