+++ Box_Smasher/Assets/Prefabs/BS/BS_Eraser_Boss.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 06bae74b4680e054092f0b1e8c352ebb
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
