+++ Box_Smasher/Assets/Prefabs/BS/BS_Eraser_Player.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 406a48824a2a9894899a23703c658f90
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
