+++ Box_Smasher/Assets/Scripts/BS/Runtime/Player/PlayerController.cs
@@ -11,9 +11,10 @@ namespace BS.Player{
 		public Animator _faceAnimator;
 		protected PlayerAnimController _animController;
 
-		Rigidbody2D _rigidBody;
+		public Rigidbody2D _rigidBody;
+		public float _maxPower = 2000; /// 최대 차지 파워
+		public float _currentPower = 0; /// 현재 차지 파워
 		Collider2D _collider;
-		public SpriteRenderer _playerBodySprite;
 		
 		public BulletEraser _eraser;
 		
@@ -26,7 +27,10 @@ namespace BS.Player{
 		public int OnHit = 0;
 		public int HP = 3;
 		public bool Invincible = false;
-		public int DEAD = 0;
+		public bool IsCharging = false;
+		public bool IsFalling = false;
+		public bool Dead = false;
+		public bool AttackSuccess = false;
 		
 		public int OnAir = 0;
 		ctw_Effector_behavior Effect;
@@ -42,20 +46,17 @@ namespace BS.Player{
 			}
 			
 			_mainCamera = CameraManager.Instance.MainCamera;
-			
 			Effect = GameObject.Find("BS_Effector").GetComponent<ctw_Effector_behavior>();
 		}
 		
 		// Maths
 		
 		float Math_2D_Force(float x, float y){
-			
 			return Mathf.Sqrt(Mathf.Pow(x,2)+Mathf.Pow(y,2));
 		}
 		
 		float Math_Boss_Damage(){
-			
-			return (500f + Mathf.Abs((_rigidBody.angularVelocity * Math_2D_Force(_rigidBody.velocity.x, _rigidBody.velocity.y) / 50f)) );
+			return (500f + Mathf.Abs((_currentPower * Math_2D_Force(_rigidBody.velocity.x, _rigidBody.velocity.y) / 50f)) );
 		}
 		
 		
@@ -86,36 +87,33 @@ namespace BS.Player{
 		}
 		
 		float Get_Angle_byPosition(Vector3 Target, Vector3 Pos){
-			
 			return ( Mathf.Atan2(Target.y-Pos.y, Target.x-Pos.x) * Mathf.Rad2Deg );
 		}
 		
 		Vector2 Get_Force_byAngle(float angle){
-			
 			return new Vector2( Mathf.Cos(angle*Mathf.Deg2Rad), Mathf.Sin(angle*Mathf.Deg2Rad) ); 
 		}
 		
 		// Checks
-		
 		void OnDamage(){
-			
 			if (HP > 1){
 				HP -= 1;
 				Invincible = true;
 				_eraser.EraserWave(0.05f);
+				_animController.OnHit();
 				Invoke("TimerInvincibleReset",1.0f);
 				CameraManager.Instance.ShakeCamera(0.5f, 0.05f);
 			}
 			else if (HP == 1){
 				HP = 0;
-				DEAD = 1;
+				Dead = true;
+				_animController.Dead();
 				CameraManager.Instance.ShakeCamera(1f, 0.1f);
 			}
 		}
 		
 		// Effects
 		void GenEffect(float angle, float F, float time, int num){
-			
 			Vector3 pos = this.transform.position;
 			Effect.Effect_Run(time, pos, Get_Force_byAngle(angle)*F, num);
 		}
@@ -125,7 +123,6 @@ namespace BS.Player{
 		// Inputs
 		
 		void InputMove(){
-			
 			float XMoveCount = 0;
 			
 			if (OnHit == 0){
@@ -159,21 +156,27 @@ namespace BS.Player{
 		void InputAttack(){
 			
 			if ((Input.GetMouseButton(0))&&(OnAir == 1)&&(OnHit != 2)){
-				if ((_rigidBody.angularVelocity <= 2000)&&(_mainCamera.ScreenToWorldPoint(Input.mousePosition).x - this.transform.position.x <= 0)){
-					_rigidBody.angularVelocity = _rigidBody.angularVelocity + 30;
+				IsCharging = true;
+				float amount = _maxPower * 0.015f;
+
+				if ((_currentPower <= _maxPower)&&(_mainCamera.ScreenToWorldPoint(Input.mousePosition).x - this.transform.position.x <= 0)){
+					_currentPower += amount;
 				}
-				if ((_rigidBody.angularVelocity >= -2000)&&(_mainCamera.ScreenToWorldPoint(Input.mousePosition).x - this.transform.position.x > 0)){
-					_rigidBody.angularVelocity = _rigidBody.angularVelocity - 30;
+				if ((_currentPower >= -_maxPower)&&(_mainCamera.ScreenToWorldPoint(Input.mousePosition).x - this.transform.position.x > 0)){
+					_currentPower -= amount;
 				}
-				if (Mathf.Abs(_rigidBody.angularVelocity) > 2000){
-					_rigidBody.angularVelocity = _rigidBody.angularVelocity/Mathf.Abs(_rigidBody.angularVelocity)*2000;
+				if (Mathf.Abs(_currentPower) > _maxPower){
+					_currentPower = _currentPower > 0 ? 1 : -1 * _maxPower;
 				}
 				OnHit = 1;
 			}
 			else if ((Input.GetMouseButtonUp(0))&&(OnHit == 1)){
-				_rigidBody.velocity = GetForceDirection()*Mathf.Abs(_rigidBody.angularVelocity)/40;
+				_rigidBody.velocity = GetForceDirection()*Mathf.Abs(_currentPower)/40;
 				OnHit = 2;
 				_collider.sharedMaterial = Bouncy;
+
+				_currentPower = 0;
+				IsCharging = false;
 			}
 			
 			if (OnHit == 1){
@@ -199,12 +202,10 @@ namespace BS.Player{
 		void Caring(){
 			
 			if ((OnAir == 1)&&(OnHit != 1)){
-				
 				_rigidBody.gravityScale = 9.8f;
 			}
 			
 			if (OnAir == 0){
-				
 				_rigidBody.gravityScale = 4.9f;
 			}
 		}
@@ -212,9 +213,7 @@ namespace BS.Player{
 		void OnTriggerStay2D(Collider2D other){
 			
 			switch (other.tag){
-				
 				case "Platform":
-				
 					ctw_Platform_behavior PlatformScript = other.GetComponent<ctw_Platform_behavior>();
 					
 					if ((PlatformScript.Trigger == false)&&(_rigidBody.velocity.y <= 0)){
@@ -223,67 +222,68 @@ namespace BS.Player{
 							GenEffect(180f, 15f, 1f, 3);
 						}
 						OnAir = 0;
+						IsFalling = false;
+						AttackSuccess = false;
 					}
-				break;
-				
+					break;
 				case "Bullet":
 					if (!Invincible){
 						ctw_Bullet_Collider_Script BulletScript = other.GetComponent<ctw_Bullet_Collider_Script>();
-						if ((DEAD != 1)&&(BulletScript.OnWork == true)) {
+						if ((!Dead)&&(BulletScript.OnWork == true)) {
 							GenEffect(Get_Angle_byPosition(this.transform.position, other.GetComponent<Transform>().position)+35f, 15f, 1f, 3);
 							GenEffect(Get_Angle_byPosition(this.transform.position, other.GetComponent<Transform>().position)-35f, 15f, 1f, 3);
 							OnDamage();
 						}
 						BulletScript.Hitted();
 					}
-				break;
-				
+					break;
 				case "Ground":
 					if ((OnAir == 1)&&(_rigidBody.velocity.y <= -1f)){
 						GenEffect(0f, 15f, 1f, 3);
 						GenEffect(180f, 15f, 1f, 3);
 					}
 					OnAir = 0;
-				break;
+					IsFalling = false;
+					AttackSuccess = false;
+					break;
 			}
 		}
 		
 		void OnTriggerExit2D(Collider2D other){
-			
 			switch (other.tag){
-				
 				case "Platform":
 					OnAir = 1;
-				break;
-				
+					IsFalling = true;
+					break;
 				case "Ground":
 					OnAir = 1;
-				break;
+					IsFalling = true;
+					break;
 			}
 		}
 		
 		void OnCollisionEnter2D(Collision2D other){
-			
-			if ((OnHit != 2)&&(other.collider.name == "BS_Boss")&&(!Invincible)){
-				if (DEAD != 1) {
+			if ( (OnHit != 2) && (other.collider.name == "BS_Boss") && (!Invincible) ){
+				if (!Dead) {
 					GenEffect(Get_Angle_byPosition(this.transform.position, other.collider.GetComponent<Transform>().position), 25f, 2f, 10);
 					OnDamage();
 				}
 			}
-			if ((OnHit == 2)&&(other.collider.name == "BS_Boss")){
-				
+			if ( (OnHit == 2) && (other.collider.name == "BS_Boss") ){
 				GenEffect(Get_Angle_byPosition(other.collider.GetComponent<Transform>().position, this.transform.position)+60f, 30f, 3f, 8);
 				GenEffect(Get_Angle_byPosition(other.collider.GetComponent<Transform>().position, this.transform.position)-60f, 30f, 3f, 8);
 
 				BossBehavior bossScript = other.collider.GetComponent<BossBehavior>();
 				bossScript.OnDamaged(Math_Boss_Damage());
+				AttackSuccess = true;
+				_animController.AttackSuccess();
 			}
 			
 			TimerAttackReset();
 		}
 		
 		void Update(){
-			if (DEAD != 1){
+			if (!Dead){
 				InputAttack();
 				InputMove();
 			}
