--- Box_Smasher/Assets/Scripts/Precoding.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 9bfcb8c127b7cc94c8e7e328131281fd
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
