+++ Box_Smasher/Packages/manifest.json
@@ -1,16 +1,16 @@
 {
   "dependencies": {
     "com.dbrizov.naughtyattributes": "https://github.com/dbrizov/NaughtyAttributes.git#upm",
-    "com.unity.2d.animation": "3.2.2",
-    "com.unity.2d.pixel-perfect": "2.0.4",
-    "com.unity.2d.psdimporter": "2.1.4",
+    "com.unity.2d.animation": "3.2.5",
+    "com.unity.2d.pixel-perfect": "2.1.0",
+    "com.unity.2d.psdimporter": "2.1.6",
     "com.unity.2d.sprite": "1.0.0",
-    "com.unity.2d.spriteshape": "3.0.11",
+    "com.unity.2d.spriteshape": "3.0.14",
     "com.unity.2d.tilemap": "1.0.0",
     "com.unity.collab-proxy": "1.2.16",
     "com.unity.ide.rider": "1.1.4",
-    "com.unity.ide.vscode": "1.2.1",
-    "com.unity.test-framework": "1.1.14",
+    "com.unity.ide.vscode": "1.2.3",
+    "com.unity.test-framework": "1.1.22",
     "com.unity.textmeshpro": "2.0.1",
     "com.unity.timeline": "1.2.15",
     "com.unity.ugui": "1.0.0",
