+++ Box_Smasher/ProjectSettings/ProjectVersion.txt
@@ -1,2 +1,2 @@
-m_EditorVersion: 2019.4.2f1
-m_EditorVersionWithRevision: 2019.4.2f1 (20b4642a3455)
+m_EditorVersion: 2019.4.20f1
+m_EditorVersionWithRevision: 2019.4.20f1 (6dd1c08eedfa)
